requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-google-genai==0.1.2
langgraph==0.0.33
chromadb==0.4.24
google-generativeai==0.3.2
tavily-python==0.0.7
python-dotenv==1.0.1